\name{plot.otolith}
\alias{plot.otolith}
\alias{plot.banana}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Visualise uncertainity in the terminal Fbar and SSB }
\description{
	Generates and visualises the two-dimensional confidence interval from a stock assessment based on a parametric bootstrapping exercise.
}
\usage{
   plot.otolith(stock, ica, plot=TRUE, show.points=FALSE, do.contours=TRUE, margin.plots=TRUE, show.estimate=TRUE,
                      xlim="missing", ylim="missing",
                      n=10000, pch=".", alpha=0.05, show.grid=TRUE
                      n.grid=50, contour.args=list(),...)
   plot.banana(stock, ica, plot=TRUE, show.points=TRUE, do.contours=FALSE, margin.plots=FALSE, show.estimate=TRUE,
                      xlim="missing", ylim="missing",
                      n=10000, pch=".", alpha=0.05, show.grid=TRUE
                      n.grid=50, contour.args=list(),...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stock}{ The FLStock object upon which the stock assessment is based.}  
  \item{ica}{ The FLICA object containing the results of the stock assessment. }
  \item{plot}{ Boolean. Whether the plot should be generated or whether or not. }
  \item{show.points}{ Boolean. Show the individual point generated by the parametric bootstrap. }
  \item{do.contours}{ Boolean. Generate contours showing the confidence intervals in the two dimensions.}
  \item{margin.plots}{ Boolean. Generate horizontal and vertical plots to the sides of the main plot showing the 1-D confidence intervals.}
  \item{show.estimate}{ Boolean. Show the point estimate generated by the assessment.}
  \item{xlim, ylim}{ Two element numeric vectors giving the axis limits for the x- and y- axes respectively.}
  \item{n}{ The number of individual parametric bootstrap values to generate.}
  \item{pch}{ The symbol used to plot the individual points. See \code{\link{points}} for more information.}
  \item{alpha}{ The confidence interval plotted in the margin plots. Defaults to 0.05 (95\%).}
  \item{show.grid}{ Boolean. Display a grid in the background.}
  \item{n.grid}{ The density of the grid used to generate the two-dimensional uncertainty surface which the contours characterise. The default value of 50 		works well for most applications, although it can be increased if the number of bootstrap estimates is also increased.}
  \item{contour.args}{ A list of arguments suitable for use in the \code{\link{contour}} function controlling the formatting of the contour lines.}
  \item{...}{ Other arguments suitable to be feed to the standard \code{\link{plot}} function.}
}
\details{
The traditional ICA function produced a function, known colloquially as the "banana plot" describing the two-dimensional correlation between the terminal Fbar and the terminal SSB, based on a parametric bootstrapping of the variance-covariance matrix from the parameter estimates. Here, we have implemented this tool in an analogous manner for use with FLICA stock assessments.

There are two distinct ways in which the results of the bootstrap can be visualised. The first is in the form of the traditional "banana-plot", where the individual bootstrap estimates are each plotted on the figure and the spread and correlation becomes visually apparent. However, the precise nature of the uncertainty surface is difficult is visualise in this manner. The second method improves upon this by calculating the uncertainty contours and plotting them directly instead, resulting in a figure akin to an otolith from a fish - hence, the name "otolith-plot". For comparison, one-dimensional uncertainty distributions are also plotted.

The input arguments to \code{plot.otolith} and \code{plot.banana} are indentical - the difference between the two functions is the default values chosen, which will generate either a "banana-plot" or an "otolith-plot" respectively. Other parameter combinations can also be selected to give a mixture of these two styles.

}
\value{
  The structure of the returned vlaue is influenced by the \code{do.contours} argument.


  \code{plot.banana}, or cases where \code{do.contours=FALSE}, returns a data-frame containing the individual boot-strap estimates of Fbar and SSB.


  \code{plot.otolith}, or cases where \code{do.contours=TRUE}, returns a list of arguments suitable for use with the \code{contour} function, that can be used later to, e.g., add extra contours to a plot.
}

\author{ Mark Payne }

\seealso{ \code{\link{sprintf}}, \code{\link{options}} }
\examples{
   #Load assessment objects
   data(herIIIa)
   data(herIIIa.tun)
   data(herIIIa.ctrl)
   #Perform assessment
   herIIIa.ica	<- FLICA(herIIIa, herIIIa.tun, herIIIa.ctrl)
   herIIIa <- herIIIa + herIIIa.ica
   #Plot otolith
   plot.otolith(herIIIa, herIIIa.ica)
   #Plot banana
   plot.banana(herIIIa, herIIIa.ica)

   #Compare the results of two assessment settings using otoliths
   herIIIa.ctrl@sep.age <- as.integer(6)		#Normally 4 in this assessment
   herIIIa.age6.ica  <- FLICA(herIIIa, herIIIa.tun, herIIIa.ctrl)
   herIIIa.age6      <- herIIIa + herIIIa.age6.ica
   #Now make the otolith plots
   plot.otolith(herIIIa, herIIIa.ica)
   age6.otolith <- plot.otolith(herIIIa.age6, herIIIa.age6.ica,plot=FALSE)
   age6.otolith$col  <- "red"
   age6.otolith$lwd  <- 2
   do.call(contour,age6.otolith)
   #Add the estimate via this method	
   points(fbar(herIIIa.age6)[,"2007"],ssb(herIIIa.age6)[,"2007"],col="red",pch=18,cex=2)
   #Finally, add a legend in the upper right
   plot.new()
   legend("topright",legend=c("Age 4","Age 6"),pch=c(19,18),col=c("black","red"),lwd=c(1,2))
}
\keyword{ classes }
